move pie chart
