+++ public/app/plugins/panel/piechart/module.tsx
@@ -1,7 +1,7 @@
 import { FieldColorModeId, FieldConfigProperty, PanelPlugin, ReducerID, standardEditorsRegistry } from '@grafana/data';
 import { PieChartPanel } from './PieChartPanel';
-import { PieChartOptions } from './types';
-import { LegendDisplayMode, PieChartType, PieChartLabels, PieChartLegendValues } from '@grafana/ui';
+import { PieChartOptions, PieChartType, PieChartLabels, PieChartLegendValues } from './types';
+import { LegendDisplayMode } from '@grafana/ui';
 import { PieChartPanelChangedHandler } from './migrations';
 import { addHideFrom } from '../timeseries/config';
